backend/requirements.txt
APScheduler==3.11.0
autograd==1.7.0
autograd-gamma==0.5.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
coverage==7.8.0
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
formulaic==1.1.1
gunicorn==23.0.0
holidays==0.68
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_resources==6.5.2
iniconfig==2.1.0
interface-meta==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lifelines==0.30.0
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pluggy==1.5.0
prophet==1.1.6
PyJWT==2.10.1
pymongo==4.11.2
pyparsing==3.2.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.38
stanio==0.5.1
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
